Bumpable now has only one bump method
Rather than lots of bump methods -- bump_spot, bump_divs etc -- Bumpable now has just one method, bump, which takes an enum of all the different types of bump. This removes a lot of the delegation boiler plate from implementations that basically just delegate it. No change to test results, as this is just a refactor.