Solvers module
Add framework for iterative solvers. Implemented implied vol solver using Brent. To get this to work I needed to change the calendar in the vol surface to be Rc rather than Box, but this is probably an improvement, as calendars are likely to be shared. (May be better still to use Arc.)