Pass val_date(s) to priceable.price method
Previously, the priceable.price method always valued as of the spot date. It can now evaluate as of multiple dates. This is particularly useful for baskets, ETFs and dynamic indices, which can show their value at any future date or dates. For example, this will make it easy to write an Asian option, or any more exotic product, on a dynamic index underlying.