Create Facade module
The facade is an easy-to-use data-driven interface to QuantMath. It can be used alongside lower-level calls. For example, the facade could be used to create an instrument, while the market data could be supplied by direct calls into the data module.

The facade is currently Rust code, but the intention is to mirror the functions in the facade with C-style or JNI interfaces, to allow interoperability with other languages.

As part of this checkin, complete the serialization and deserialization of all objects required for pricing.