Thread safety
Added Send + Sync to all the objects that are passed through the external interface, as we have no way of policing whether they are used in different threads. Fix all the compile errors that result.

The effect is that all Rc containers are now Arc. This makes sense, as we really only use sharing when external users want it. For example, a user may use the same instrument and market data in different threads. It adds an overhead to cloning an Arc, but we do not do that very often.